fix(actualserver): update chart with the right image (#15629)
**Description**
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->

The
[`actualserver`](https://github.com/truecharts/charts/tree/master/charts/stable/actualserver)
chart is using [this
image](https://hub.docker.com/r/jlongster/actual-server), but releases
are now done
[here](https://github.com/actualbudget/actual-server/pkgs/container/actual-server)
and [here](https://hub.docker.com/r/actualbudget/actual-server). As
such, this PR updates the image to the correct one, as well as updating
the app version.

⚒️ Fixes  #15628

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [x] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

I'm honestly unsure how to test chart modifications, but if needed, I'll
figure out a way. I do know the new image is correct, as I have deployed
it on my local machine for testing.

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [x] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning