fix(notifiarr): Fix inability to change auth settings (#15294)
**Description**
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->
Notifiarr treats env as a higher priority than its config file.
Therefore, by setting the `DN_UI_PASSWORD` the user is not able to
change the password or auth method via the Notifiarr UI.

However, Notifiarr also generates a random password on the first run,
which is stored in its config file. Therefore, if the `DB_UI_PASSWORD`
env is not set the user will be expected to input the randomly generated
password. To obtain it, the user must remember to inspect the logs on
the first run of the app.

To avoid this, I have added a simple init container that creates an
empty config file (containing only a comment). This prevents Notifiarr
from generating a new config file and thus a random password.

When the config file is present but no password is set in it, Notifiarr
accepts logins with the username of `admin` and the value of the
`DN_API_KEY` env as the password.

⚒️ Fixes  # <!--(issue)-->

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [X] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->
Deployed via helm and tried to log in with the set API key. Then, I
tried setting a password and changing the auth method.

**📃 Notes:**
<!-- Please enter any other relevant information here -->
Note: Now that the `DB_UI_PASSWORD` env is not set anymore, the will be
expected to log in with the password that is stored in the config file.
If they haven't tried to change it via the UI, that password will be the
randomly generated one on the first run of the app.

Because of this, I am wondering if we should flag this update as
breaking instead of as just a fix?

**✔️ Checklist:**

- [X] ⚖️ My code follows the style guidelines of this project
- [X] 👀 I have performed a self-review of my own code
- [X] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [X] 📄 I have made corresponding changes to the documentation
- [X] ⚠️ My changes generate no new warnings
- [X] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [X] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🪞 I have opened a PR on
[truecharts/containers](https://github.com/truecharts/containers) adding
the container to TrueCharts mirror repo.
- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: sdimovv <36302090+sdimovv@users.noreply.github.com>
Signed-off-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>
Co-authored-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>