feat(bazarr) Adding exportarr to expose metrics to prometheus (#14761)
**Description**
This adds exportarr to the bazarr application as bazarr is supported by
exportarr. This allows the user to have prometheus scrape for bazarr
metrics.

**⚙️ Type of change**

- [X] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
Created a personal catalog which I added to TrueNAS. I then installed
this app sourced from this catalog to ensure the questions + answers,
and actual installation works.

**📃 Notes:**
Exportarr for bazarr does not work without having the API key to access
bazarr set. I have added this as a question to the metrics part (see
below screenshot). Is this an acceptable way of doing this, or should it
be done differently? Unfortunately bazarr does not support setting the
api key via environment variables, so enabling metrics for a new
deployment requires 2 deployments. The first one without metrics enabled
to bootstrap bazarr, and then a second one to enable metrics and set the
api key.


![2023-11-20_12-26](https://github.com/truecharts/charts/assets/20680059/b9455821-6dc5-4f9e-bc8f-16461bbb394d)

Also, due to a bug in exportarr
(https://github.com/onedr0p/exportarr/issues/244), exportarr does not
work when there are no series added/sonarr is not configured. Should
this be mentioned anywhere to the users of this chart?

**✔️ Checklist:**

- [X] ⚖️ My code follows the style guidelines of this project
- [X] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [X] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🪞 I have opened a PR on
[truecharts/containers](https://github.com/truecharts/containers) adding
the container to TrueCharts mirror repo.
- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: Kjeld Schouten <kjeld@schouten-lebbing.nl>
Co-authored-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>
Co-authored-by: Kjeld Schouten <kjeld@schouten-lebbing.nl>