chore(nextcloud): BREAKING CHANGE remove service settings for SCALE GUI (#16379)
**Description**

Users seem to be confused about the fact service settings where present
and even defaulted to loadbalancer, without it actually being
technically possible to use this without causing malfuctions.

Requiring ingress is a good first step, but in the specific case of
Nextcloud, we actually want to prevent any use of direct ip:port
connections to nextcloud on SCALE.

It won't work nicely without users writhing custom forwarding scripts
which 99.99% of users on SCALE isn't able to do and those that do really
should just use normal helm.

Also removes podOptions as those options really shouldn't be used on
Nextcloud either.

As well as ServiceExpert, as we definately don't want direct interface
attachment on this either.

Only breaking for users currently using IP-Port connections on TrueNAS
SCALE, which they shouldn't have been doing in the first place

**⚙️ Type of change**

- [x] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [x] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

**📃 Notes:**
<!-- Please enter any other relevant information here -->

**✔️ Checklist:**

- [ ] ⚖️ My code follows the style guidelines of this project
- [ ] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [ ] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🪞 I have opened a PR on
[truecharts/containers](https://github.com/truecharts/containers) adding
the container to TrueCharts mirror repo.
- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: Kjeld Schouten <kjeld@schouten-lebbing.nl>