fix(apps): fix unpackerr config storage (#16826)
**Description**
Fixes an issue where the Unpackerr chart would not automatically create
(or load) unpackerr.conf unless the directory (/config) was created and
mapped. The /downloads mapping was removed and replaced with /config
instead. The chart will start without the directory present, however, it
will only rely on environment variables.

⚒️ Fixes:  
- Maps /config so that the app can create and read unpackerr.conf
instead of relying on environment variables.
- Added whisparr to chart.yaml since it is supported by Unpackerr

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [X] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
Make changes to affected files and start the chart.

**✔️ Checklist:**

- [X] ⚖️ My code follows the style guidelines of this project
- [X] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [X] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [X] ⬆️ I increased versions for any altered app according to semantic
versioning