feat(minecraft-java) add wider packwiz support (#17523)
**Description**

Updated minecraft-java's questions.yaml to allow downloading packwiz
packs when modloaders other than Fabric are chosen. The developer of the
docker image has
[stated](https://github.com/itzg/docker-minecraft-server/issues/1238#issuecomment-1003560897)
that packwiz support is not limited to Fabric. Mod loaders supported by
packwiz itself are listed
[here](https://packwiz.infra.link/reference/commands/packwiz/init/#options).

**⚙️ Type of change**

- [x] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**

I don't have the capability to test this at this time.

**📃 Notes:**

I haven't worked on TrueNAS apps/questions.yaml before but based on my
understanding from reading the docs this should be correct. If there is
a way to combine all the added lines using a compound show-if statement
(ie show-if type is forge or fabric or quilt or neoforge) I couldn't
find documentation on it.

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [ ] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [ ] ⬆️ I increased versions for any altered app according to semantic
versioning

---------

Signed-off-by: H <37481947+H-QueerCoded@users.noreply.github.com>