feat(speedtest-exporter) permission fix and extending timeout
with speed tests, they can take more than the defaulted time of 10 seconds. I have made the speed test 60 seconds. A second issue is that the add is not being ran by admin so you are unable to modify necessary files that require the app to run.

adding chart change