Add option to enable Remote Write Receiver in Prometheus (#17739)
**Description**
Prometheus has the ability to receive metrics, rather then pull them.
Ref: https://prometheus.io/docs/prometheus/latest/storage/#overview
The current chart has the required parameters to make the Prometheus
instance a sender, but not a receiver. This PR adds an option to enable
the Remote Write Receiver functionality. The flag is provided in the
prometheus-operator CRD, so all we have to do is add the question,
values, and update the template. This PR attempts to do that.

**⚙️ Type of change**

- [x] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
Tested changes on local TrueNAS

**📃 Notes:**
None

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning

---

Signed-off-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>
Co-authored-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>