feat(cloudreve) Migrate from unmaintained to upstream images. (#17489)
**Description**
Replaces the 2-year-outdated cloudreve image with upstream v.3.8.3,
which includes aria2 support and English translations.
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->
⚒️ Fixes  #16912 <!--(issue)-->

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [x] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
Tested for page load, login, Aria2 setup locally with k8s + Docker
Desktop. Have not tested any downloads or real use.
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

**📃 Notes:**

- Have not tested beyond initial "it runs and I can log in". Should I be
copying this app into dev or incubator?
- Upstream docker image does not appear to support passing RPC_SECRET,
which leads to a mess chasing down the auto-generated secret and
manually entering it on the webpage. Should this be a questions.yaml
entry instead?
- Since cloudreve requires configuration to setup and use aria2 anyway,
should the aria2 image be removed and documentation added to refer to
the aria2-app (https://truecharts.org/charts/stable/aria2/)?

<!-- Please enter any other relevant information here -->

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [x] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: Christopher <bigbero@gmail.com>
Co-authored-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>